test: StaticBetaRndの乱数テストをクラスに分離
