+++ test/matsu/num/statistics/random/staticbeta/GammaBasedStaticBetaRndTest.java
@@ -10,8 +10,6 @@ import static matsu.num.statistics.random.StaticBetaRnd.*;
 import static org.hamcrest.MatcherAssert.*;
 import static org.hamcrest.Matchers.*;
 
-import java.util.Objects;
-
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.Test.None;
@@ -21,7 +19,6 @@ import org.junit.runner.RunWith;
 import matsu.num.statistics.random.BaseRandom;
 import matsu.num.statistics.random.FloatingRandomGeneratorTestingFramework;
 import matsu.num.statistics.random.StaticBetaRnd;
-import matsu.num.statistics.random.TestedFloatingRandomGenerator;
 import matsu.num.statistics.random.staticgamma.StaticGammaFactoryForTesting;
 
 /**
@@ -65,92 +62,21 @@ final class GammaBasedStaticBetaRndTest {
 
     public static class ベータ乱数のテスト {
 
-        private FloatingRandomGeneratorTestingFramework framework;
-
-        @Before
-        public void before() {
-            framework = FloatingRandomGeneratorTestingFramework
-                    .instanceOf(new TestedBeta(BaseRandom.threadSeparatedRandom()));
-        }
-
         @Test
         public void test() {
-            framework.test();
-        }
-
-        /**
-         * a=2,b=1のベータ乱数発生器のテスタ. <br>
-         * 累積分布関数はx^2
-         */
-        private static final class TestedBeta implements TestedFloatingRandomGenerator {
-
-            private final BaseRandom random;
-            private final double a;
-            private final double b;
-            private final StaticBetaRnd staticBetaRnd = FACTORY.instance();
-
-            public TestedBeta(BaseRandom random) {
-                this.random = Objects.requireNonNull(random);
-                a = 2;
-                b = 1;
-            }
-
-            @Override
-            public double newValue() {
-                return this.staticBetaRnd.nextRandom(random, a, b);
-            }
-
-            @Override
-            public double cumulativeProbability(double arg) {
-                return arg * arg;
-            }
-
+            FloatingRandomGeneratorTestingFramework
+                    .instanceOf(new Tested_At_2_1_StaticBetaRandomGenerator(FACTORY.instance()))
+                    .test();
         }
     }
 
     public static class ベータプライム乱数のテスト {
 
-        private FloatingRandomGeneratorTestingFramework framework;
-
-        @Before
-        public void before() {
-            framework = FloatingRandomGeneratorTestingFramework
-                    .instanceOf(new TestedBetaPrime(BaseRandom.threadSeparatedRandom()));
-        }
-
         @Test
         public void test() {
-            framework.test();
-        }
-
-        /**
-         * a=2,b=1のベータプライム乱数発生器のテスタ. <br>
-         * 累積分布関数は<br>
-         * {@literal 1 - (1+2x)/(1+x)^2}
-         */
-        private static final class TestedBetaPrime implements TestedFloatingRandomGenerator {
-
-            private final BaseRandom random;
-            private final double a;
-            private final double b;
-            private final StaticBetaRnd staticBetaRnd = FACTORY.instance();
-
-            public TestedBetaPrime(BaseRandom random) {
-                this.random = Objects.requireNonNull(random);
-                a = 2;
-                b = 1;
-            }
-
-            @Override
-            public double newValue() {
-                return staticBetaRnd.nextBetaPrime(random, a, b);
-            }
-
-            @Override
-            public double cumulativeProbability(double arg) {
-                return 1 - (1 + 2 * arg) / ((1 + arg) * (1 + arg));
-            }
-
+            FloatingRandomGeneratorTestingFramework
+                    .instanceOf(new Tested_At_2_1_StaticBetaPrimeRandomGenerator(FACTORY.instance()))
+                    .test();
         }
     }
